Addressed a few more TODOs; includes updates to how TrainingOutput gets produced
