+++ test/test_reports.py
@@ -196,13 +196,13 @@ def make_dummy_trained_model(index: int, tmp_path, accuracy_value: float = 0.5, 
             "batch_size": 0,
             "epochs": 0,
             "model_architecture": {},
+            "model_name": "",
             "seed": 0,
             "training_dataset_config_path": "",
         },
         "results": {
             "accuracy": [],
             "loss": [],
-            "model_name": "",
             "val_accuracy": [],
             "val_loss": [],
         },
